refactor(useboard): implement useBoard part 1
